fix(frontend): preserving auth / permission states after hot-reload
