+++ packages/frontend/src/contexts/auth.context.tsx
@@ -4,24 +4,10 @@
  * Full terms: see LICENSE.md.
  */
 
-import { createContext, ReactNode } from "react";
+import { createContext } from "react";
 
-import { Progress } from "@/app-components/displays/Progress";
-import { runtimeConfig } from "@/config/runtime";
-import {
-  useTanstackQuery,
-  useTanstackQueryClient,
-} from "@/hooks/crud/useTanstack";
-import { useLogout } from "@/hooks/entities/auth-hooks";
-import { useApiClient } from "@/hooks/useApiClient";
-import { useAppRouter } from "@/hooks/useAppRouter";
-import { CURRENT_USER_KEY } from "@/hooks/useAuth";
-import { useSubscribeBroadcastChannel } from "@/hooks/useSubscribeBroadcastChannel";
-import { useTranslate } from "@/hooks/useTranslate";
-import { RouterType } from "@/services/types";
 import { UseMutateFunction } from "@/types/tanstack.types";
 import { IUser } from "@/types/user.types";
-import { hasPublicPath, isLoginPath } from "@/utils/URL";
 
 export interface AuthContextValue {
   user: IUser | undefined;
@@ -42,87 +28,3 @@ export const AuthContext = createContext<AuthContextValue>({
 });
 
 AuthContext.displayName = "AuthContext";
-
-export interface AuthProviderProps {
-  children: ReactNode;
-}
-
-export const AuthProvider = ({ children }: AuthProviderProps): JSX.Element => {
-  const router = useAppRouter();
-  const { i18n } = useTranslate();
-  const queryClient = useTanstackQueryClient();
-  const updateLanguage = (lang: string) => {
-    i18n.changeLanguage(lang);
-  };
-  const { mutate: logoutSession } = useLogout();
-  const logout = async () => {
-    updateLanguage(runtimeConfig.lang.default);
-    logoutSession();
-  };
-  const authRedirection = async (isAuthenticated: boolean) => {
-    if (isAuthenticated && router.routeObject.handle?.isPublicRoute) {
-      await router.push(RouterType.HOME);
-    }
-
-    if (isAuthenticated && isLoginPath(router.pathname)) {
-      const rawRedirect = router.query.redirect;
-      const redirectUrl = Array.isArray(rawRedirect)
-        ? rawRedirect.at(-1)
-        : rawRedirect;
-
-      if (redirectUrl?.startsWith("/") && !hasPublicPath(redirectUrl)) {
-        await router.push(redirectUrl);
-      } else {
-        await router.push(RouterType.HOME);
-      }
-    }
-  };
-  const { apiClient } = useApiClient();
-  const {
-    data: user,
-    error,
-    isLoading,
-  } = useTanstackQuery<IUser, Error>({
-    queryFn: () => apiClient.getCurrentSession(),
-    queryKey: [CURRENT_USER_KEY],
-    onSuccess: (data) => {
-      updateLanguage(data.language);
-      authRedirection(!!data.id);
-    },
-  });
-  const setUser = (data?: IUser) => {
-    queryClient.setQueryData([CURRENT_USER_KEY], data);
-  };
-  const authenticate = (user: IUser) => {
-    updateLanguage(user.language);
-    setUser(user);
-  };
-  const isAuthenticated = !!user;
-
-  useSubscribeBroadcastChannel("login", () => {
-    router.reload();
-  });
-
-  useSubscribeBroadcastChannel("logout", () => {
-    router.reload();
-  });
-
-  if (isLoading) {
-    return <Progress />;
-  }
-
-  return (
-    <AuthContext.Provider
-      value={{
-        user,
-        isAuthenticated: !!isAuthenticated,
-        error,
-        setUser,
-        authenticate,
-        logout,
-      }}
-    >
-      {children}
-    </AuthContext.Provider>
-  );
-};
