+++ packages/frontend/src/contexts/permission.context.tsx
@@ -4,10 +4,8 @@
  * Full terms: see LICENSE.md.
  */
 
-import { createContext, ReactNode, useCallback, useMemo } from "react";
+import { createContext } from "react";
 
-import { Progress } from "@/app-components/displays/Progress";
-import { useUserPermissions } from "@/hooks/entities/auth-hooks";
 import { EntityType } from "@/services/types";
 import { PermissionAction } from "@/types/permission.types";
 
@@ -18,47 +16,3 @@ export const PermissionContext = createContext<{
 });
 
 PermissionContext.displayName = "PermissionContext";
-
-export interface PermissionProviderProps {
-  children: ReactNode;
-}
-
-export const PermissionProvider = ({
-  children,
-}: PermissionProviderProps): JSX.Element => {
-  const { data, isLoading, isFetching } = useUserPermissions();
-  const permissionMap = useMemo(
-    () =>
-      data?.permissions.reduce(
-        (acc, { model, action }) => {
-          if (!(model in acc)) {
-            acc[model] = [];
-          }
-
-          acc[model]?.push(action);
-
-          return acc;
-        },
-        {} as { [key in EntityType]?: PermissionAction[] },
-      ),
-    [data?.permissions],
-  );
-  const getAllowedActions = useCallback(
-    (type: EntityType) => {
-      return permissionMap && type in permissionMap ? permissionMap[type] : [];
-    },
-    [permissionMap],
-  );
-
-  if (isLoading || isFetching) return <Progress />;
-
-  return (
-    <PermissionContext.Provider
-      value={{
-        getAllowedActions,
-      }}
-    >
-      {children}
-    </PermissionContext.Provider>
-  );
-};
